mv CollectionArtifactFile to models.models
