--- ansible_galaxy/models/collection_artifact_file.py
@@ -1,19 +1,0 @@
-
-import logging
-
-import attr
-
-# from ansible_galaxy.models.collection_info import CollectionInfo
-log = logging.getLogger(__name__)
-
-
-# see https://github.com/ansible/galaxy/issues/957
-@attr.s(frozen=True)
-class CollectionArtifactFile(object):
-    name = attr.ib()
-    ftype = attr.ib()
-    src_name = attr.ib(default=None)
-    chksum_type = attr.ib(default="sha256")
-    chksum_sha256 = attr.ib(default=None)
-    # name = attr.ib(default=None)
-    format_version = attr.ib(default=0.0)
