+++ ansible_galaxy/models/models.py
@@ -9,3 +9,15 @@ log = logging.getLogger(__name__)
 class BuildContext(object):
     collection_path = attr.ib()
     output_path = attr.ib()
+
+
+# see https://github.com/ansible/galaxy/issues/957
+@attr.s(frozen=True)
+class CollectionArtifactFile(object):
+    name = attr.ib()
+    ftype = attr.ib()
+    src_name = attr.ib(default=None)
+    chksum_type = attr.ib(default="sha256")
+    chksum_sha256 = attr.ib(default=None)
+    # name = attr.ib(default=None)
+    format_version = attr.ib(default=0.0)
